% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{backend_needs_express}
\alias{backend_needs_express}
\title{Check if a backend requires Express dependencies}
\usage{
backend_needs_express(backend_module)
}
\arguments{
\item{backend_module}{Character string. The backend module filename.}
}
\value{
Logical.
}
\description{
Check if a backend requires Express dependencies
}
\keyword{internal}
