% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_npm_path}
\alias{cache_npm_path}
\title{Get path to npm packages cache}
\usage{
cache_npm_path()
}
\value{
Character string. The path to the npm packages cache.
}
\description{
Determines the path to the cache directory for npm packages used by shinyelectron.
}
\section{Details}{

The npm packages are cached at \code{cache_dir()/npm}. This allows for reuse
of downloaded npm dependencies across multiple builds.
}

\keyword{internal}
