% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{CONFIG_FILENAME}
\alias{CONFIG_FILENAME}
\title{Configuration file name}
\format{
An object of class \code{character} of length 1.
}
\usage{
CONFIG_FILENAME
}
\description{
Configuration file name
}
\keyword{internal}
