% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-templates.R
\name{copy_backend_modules}
\alias{copy_backend_modules}
\title{Copy backend module(s) and their shared JS helpers into the build}
\usage{
copy_backend_modules(output_dir, backend_module, is_multi_app)
}
\description{
Copy backend module(s) and their shared JS helpers into the build
}
\keyword{internal}
