% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\title{Dependency Detection and Installation}
\description{
Functions for detecting, merging, and installing R and Python package
dependencies for native Shiny app types.
}
\keyword{internal}
