% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_current_arch}
\alias{detect_current_arch}
\title{Detect current architecture}
\usage{
detect_current_arch()
}
\value{
Character string representing current architecture ("x64" or "arm64")
}
\description{
Detect current architecture
}
\keyword{internal}
