% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_current_platform}
\alias{detect_current_platform}
\title{Detect current platform}
\usage{
detect_current_platform()
}
\value{
Character string representing current platform ("win", "mac", or "linux")
}
\description{
Detect current platform
}
\keyword{internal}
