% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{find_config}
\alias{find_config}
\title{Find configuration file}
\usage{
find_config(appdir)
}
\arguments{
\item{appdir}{Character path to app directory}
}
\value{
Character path to config file, or NULL if not found
}
\description{
Searches for _shinyelectron.yml in the given directory.
}
\keyword{internal}
