% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-python.R
\name{find_python_command}
\alias{find_python_command}
\title{Find the Python command}
\usage{
find_python_command()
}
\value{
Character string or NULL. The Python command name, or NULL if not found.
}
\description{
Searches for python3 first, then python on the system PATH.
}
\keyword{internal}
