% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-python.R
\name{generate_python_runtime_manifest}
\alias{generate_python_runtime_manifest}
\title{Generate a Python runtime manifest for auto-download}
\usage{
generate_python_runtime_manifest(
  version,
  platform = NULL,
  arch = NULL,
  release_date = NULL
)
}
\arguments{
\item{version}{Character string. Python version.}

\item{platform}{Character string. Target platform.}

\item{arch}{Character string. Target architecture.}

\item{release_date}{Character string. python-build-standalone release tag
(YYYYMMDD). If NULL, resolved automatically via \code{python_resolve_pbs()}.}
}
\value{
Character string. JSON content.
}
\description{
Generate a Python runtime manifest for auto-download
}
\keyword{internal}
