% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_default}
\alias{get_default}
\title{Get a default value}
\usage{
get_default(key, default = NULL)
}
\arguments{
\item{key}{Character name of the default to retrieve}

\item{default}{Value to return if key not found}
}
\value{
The default value
}
\description{
Retrieves a default value from SHINYELECTRON_DEFAULTS.
}
\keyword{internal}
