% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-metadata.R
\name{is_http_url}
\alias{is_http_url}
\title{Test for an http or https URL}
\usage{
is_http_url(x)
}
\arguments{
\item{x}{Value to test.}
}
\value{
\code{TRUE} when \code{x} is a single string that starts with \verb{http://} or
\verb{https://} (in any case) followed by more text, otherwise \code{FALSE}.
}
\description{
Test for an http or https URL
}
\keyword{internal}
