% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-app-type.R
\name{is_multi_app}
\alias{is_multi_app}
\title{Check if config is multi-app mode}
\usage{
is_multi_app(config)
}
\arguments{
\item{config}{List. Configuration.}
}
\value{
Logical.
}
\description{
Check if config is multi-app mode
}
\keyword{internal}
