% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{merge_config_deep}
\alias{merge_config_deep}
\title{Deep merge two lists}
\usage{
merge_config_deep(defaults, config)
}
\arguments{
\item{defaults}{List of default values}

\item{config}{List of config values to merge}
}
\value{
Merged list
}
\description{
Recursively merges config into defaults, where config values override defaults.
}
\keyword{internal}
