% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodejs.R
\name{nodejs_download_checksums}
\alias{nodejs_download_checksums}
\title{Download Node.js checksums}
\usage{
nodejs_download_checksums(version)
}
\arguments{
\item{version}{Character Node.js version}
}
\value{
Named character vector (filename = checksum)
}
\description{
Download Node.js checksums
}
\keyword{internal}
