% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodejs.R
\name{nodejs_download_url}
\alias{nodejs_download_url}
\title{Generate Node.js download URL}
\usage{
nodejs_download_url(version, platform, arch)
}
\arguments{
\item{version}{Character Node.js version (e.g., "22.0.0")}

\item{platform}{Character platform ("win", "mac", "linux")}

\item{arch}{Character architecture ("x64", "arm64")}
}
\value{
Character URL to download Node.js
}
\description{
Generate Node.js download URL
}
\keyword{internal}
