% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-nodejs.R
\name{nodejs_executable}
\alias{nodejs_executable}
\title{Get path to local Node.js executable}
\usage{
nodejs_executable(version = NULL)
}
\arguments{
\item{version}{Character Node.js version (NULL = latest installed)}
}
\value{
Character path to node executable, or NULL if not found
}
\description{
Get path to local Node.js executable
}
\keyword{internal}
