% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodejs.R
\name{nodejs_install_path}
\alias{nodejs_install_path}
\title{Get Node.js installation path in cache}
\usage{
nodejs_install_path(version = NULL, platform = NULL, arch = NULL)
}
\arguments{
\item{version}{Character Node.js version (NULL for base path)}

\item{platform}{Character platform (defaults to current)}

\item{arch}{Character architecture (defaults to current)}
}
\value{
Character path to Node.js installation
}
\description{
Get Node.js installation path in cache
}
\keyword{internal}
