% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodejs.R
\name{nodejs_is_installed}
\alias{nodejs_is_installed}
\title{Check if Node.js is installed locally}
\usage{
nodejs_is_installed(version = NULL)
}
\arguments{
\item{version}{Character Node.js version (NULL = any version)}
}
\value{
Logical TRUE if installed
}
\description{
Check if Node.js is installed locally
}
\keyword{internal}
