% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-nodejs.R
\name{nodejs_list_installed}
\alias{nodejs_list_installed}
\title{List installed Node.js versions}
\usage{
nodejs_list_installed()
}
\value{
Character vector of installed versions (sorted newest first)
}
\description{
List installed Node.js versions
}
\keyword{internal}
