% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodejs.R
\name{nodejs_remove}
\alias{nodejs_remove}
\title{Remove Node.js installation}
\usage{
nodejs_remove(version = NULL, verbose = TRUE)
}
\arguments{
\item{version}{Character version to remove. If NULL, removes all versions.}

\item{verbose}{Logical whether to show progress. Default TRUE.}
}
\value{
Invisible NULL
}
\description{
Remove Node.js installation
}
\keyword{internal}
