% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-nodejs.R
\name{nodejs_shasums_url}
\alias{nodejs_shasums_url}
\title{Get Node.js SHASUMS URL}
\usage{
nodejs_shasums_url(version)
}
\arguments{
\item{version}{Character Node.js version}
}
\value{
Character URL to SHASUMS256.txt
}
\description{
Get Node.js SHASUMS URL
}
\keyword{internal}
