% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodejs.R
\name{npm_executable}
\alias{npm_executable}
\title{Get path to local npm executable}
\usage{
npm_executable(version = NULL)
}
\arguments{
\item{version}{Character Node.js version (NULL = latest installed)}
}
\value{
Character path to npm executable, or NULL if not found
}
\description{
Get path to local npm executable
}
\keyword{internal}
