% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{parse_requirements_txt}
\alias{parse_requirements_txt}
\title{Parse requirements.txt file}
\usage{
parse_requirements_txt(path)
}
\arguments{
\item{path}{Character string. Path to requirements.txt.}
}
\value{
Character vector of package names.
}
\description{
Parse requirements.txt file
}
\keyword{internal}
