% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{python_executable}
\alias{python_executable}
\title{Get the path to the Python executable in a cached installation}
\usage{
python_executable(version, platform = NULL, arch = NULL)
}
\arguments{
\item{version}{Character string. Python version.}

\item{platform}{Character string. Platform (default: current).}

\item{arch}{Character string. Architecture (default: current).}
}
\value{
Character string or NULL.
}
\description{
Get the path to the Python executable in a cached installation
}
\keyword{internal}
