% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{python_install_path}
\alias{python_install_path}
\title{Get the installation path for a cached Python version}
\usage{
python_install_path(version, platform = NULL, arch = NULL)
}
\arguments{
\item{version}{Character string. Python version.}

\item{platform}{Character string. Platform (default: current).}

\item{arch}{Character string. Architecture (default: current).}
}
\value{
Character string. Path to the cached Python installation.
}
\description{
Get the installation path for a cached Python version
}
\keyword{internal}
