% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-python.R
\name{python_sha256sums_url}
\alias{python_sha256sums_url}
\title{URL of the python-build-standalone SHA256SUMS file for a release}
\usage{
python_sha256sums_url(release_date)
}
\arguments{
\item{release_date}{Character. python-build-standalone release tag (YYYYMMDD).}
}
\value{
Character URL.
}
\description{
URL of the python-build-standalone SHA256SUMS file for a release
}
\keyword{internal}
