% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-r.R
\name{r_executable}
\alias{r_executable}
\title{Get the path to the Rscript executable in a cached installation}
\usage{
r_executable(version, platform = NULL, arch = NULL)
}
\arguments{
\item{version}{Character string. R version.}

\item{platform}{Character string. Platform (default: current).}

\item{arch}{Character string. Architecture (default: current).}
}
\value{
Character string or NULL. Path to Rscript, or NULL if not found.
}
\description{
Get the path to the Rscript executable in a cached installation
}
\keyword{internal}
