% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-r.R
\name{r_install_path}
\alias{r_install_path}
\title{Get the installation path for a cached R version}
\usage{
r_install_path(version, platform = NULL, arch = NULL)
}
\arguments{
\item{version}{Character string. R version.}

\item{platform}{Character string. Platform (default: current).}

\item{arch}{Character string. Architecture (default: current).}
}
\value{
Character string. Path to the cached R installation.
}
\description{
Get the installation path for a cached R version
}
\keyword{internal}
