% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{r_is_installed}
\alias{r_is_installed}
\title{Check if a portable R version is installed}
\usage{
r_is_installed(version, platform = NULL, arch = NULL)
}
\arguments{
\item{version}{Character string. R version to check.}

\item{platform}{Character string. Platform (default: current).}

\item{arch}{Character string. Architecture (default: current).}
}
\value{
Logical.
}
\description{
Check if a portable R version is installed
}
\keyword{internal}
