% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read configuration file}
\usage{
read_config(appdir)
}
\arguments{
\item{appdir}{Character path to app directory}
}
\value{
List of configuration values (merged with defaults)
}
\description{
Reads and parses _shinyelectron.yml from the app directory.
If the file doesn't exist, returns default configuration.
}
\keyword{internal}
