% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-templates.R
\name{render_shared_templates}
\alias{render_shared_templates}
\title{Render shared Electron templates (main.js, lifecycle.html, preload.js, launcher.html)}
\usage{
render_shared_templates(output_dir, template_vars, is_multi_app)
}
\description{
Render shared Electron templates (main.js, lifecycle.html, preload.js, launcher.html)
}
\keyword{internal}
