% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restore_environment}
\alias{restore_environment}
\title{Restore environment variables}
\usage{
restore_environment(old_env)
}
\arguments{
\item{old_env}{Named list of environment variables to restore}
}
\description{
Restore environment variables
}
\keyword{internal}
