% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_dev_environment}
\alias{set_dev_environment}
\title{Set development environment variables}
\usage{
set_dev_environment(port, open_devtools)
}
\arguments{
\item{port}{Integer port number}

\item{open_devtools}{Logical whether to open dev tools}
}
\value{
Named list of old environment variables
}
\description{
Set development environment variables
}
\keyword{internal}
