% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_electron_project}
\alias{setup_electron_project}
\title{Setup Electron project structure}
\usage{
setup_electron_project(output_dir, app_name, app_type, verbose = TRUE)
}
\arguments{
\item{output_dir}{Character path to output directory}

\item{app_name}{Character application name}

\item{app_type}{Character application type}

\item{verbose}{Logical whether to show progress}
}
\description{
Setup Electron project structure
}
\keyword{internal}
