% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{sitrep}
\alias{sitrep}
\title{Diagnostic Functions for shinyelectron}
\description{
These functions provide diagnostic information about your system setup
and help troubleshoot common issues with shinyelectron.
}
\keyword{internal}
