% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_app_type}
\alias{validate_app_type}
\title{Validate application type}
\usage{
validate_app_type(app_type)
}
\arguments{
\item{app_type}{Character application type}
}
\description{
Validate application type
}
\keyword{internal}
