% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_arch}
\alias{validate_arch}
\title{Validate target architecture}
\usage{
validate_arch(arch)
}
\arguments{
\item{arch}{Character vector of architectures}
}
\description{
Validate target architecture
}
\keyword{internal}
