% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_build_output}
\alias{validate_build_output}
\title{Validate build output}
\usage{
validate_build_output(output_dir, platform)
}
\arguments{
\item{output_dir}{Character Electron project directory}

\item{platform}{Character vector of target platforms}
}
\description{
Validate build output
}
\keyword{internal}
