% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_container_available}
\alias{validate_container_available}
\title{Validate a container engine is available}
\usage{
validate_container_available(preference = NULL)
}
\arguments{
\item{preference}{Character string or NULL. Preferred engine.}
}
\value{
Invisible character string with the engine name.
}
\description{
Checks that Docker or Podman is installed and can be executed.
}
\keyword{internal}
