% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_directory_exists}
\alias{validate_directory_exists}
\title{Validate that a directory exists}
\usage{
validate_directory_exists(dir, name = "Directory")
}
\arguments{
\item{dir}{Character path to directory}

\item{name}{Character descriptive name for error messages}
}
\description{
Validate that a directory exists
}
\keyword{internal}
