% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_electron_project}
\alias{validate_electron_project}
\title{Validate Electron project structure}
\usage{
validate_electron_project(app_dir)
}
\arguments{
\item{app_dir}{Character path to Electron app directory}
}
\description{
Validate Electron project structure
}
\keyword{internal}
