% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_platform}
\alias{validate_platform}
\title{Validate target platform}
\usage{
validate_platform(platform)
}
\arguments{
\item{platform}{Character vector of platforms}
}
\description{
Validate target platform
}
\keyword{internal}
