% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-app-structure.R
\name{validate_python_app_structure}
\alias{validate_python_app_structure}
\title{Validate Python app structure}
\usage{
validate_python_app_structure(appdir)
}
\arguments{
\item{appdir}{Character string. Path to the app directory.}
}
\description{
Validate Python app structure
}
\keyword{internal}
