% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_python_available}
\alias{validate_python_available}
\title{Validate Python is available on the system}
\usage{
validate_python_available()
}
\value{
Invisible TRUE if Python is found, otherwise aborts.
}
\description{
Validate Python is available on the system
}
\keyword{internal}
