% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-app-type.R
\name{validate_runtime_strategy}
\alias{validate_runtime_strategy}
\title{Validate runtime strategy}
\usage{
validate_runtime_strategy(strategy)
}
\arguments{
\item{strategy}{Character string. The runtime strategy to validate.}
}
\description{
Validate runtime strategy
}
\keyword{internal}
