% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-app-type.R
\name{validate_runtime_strategy_for_app_type}
\alias{validate_runtime_strategy_for_app_type}
\title{Validate runtime strategy is compatible with app type}
\usage{
validate_runtime_strategy_for_app_type(strategy, app_type)
}
\arguments{
\item{strategy}{Character string. The runtime strategy.}

\item{app_type}{Character string. The app type.}
}
\description{
Validate runtime strategy is compatible with app type
}
\keyword{internal}
