% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_shiny_app_structure}
\alias{validate_shiny_app_structure}
\title{Validate Shiny application structure}
\usage{
validate_shiny_app_structure(appdir)
}
\arguments{
\item{appdir}{Character path to Shiny app directory}
}
\description{
Validate Shiny application structure
}
\keyword{internal}
