% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-stages.R
\name{write_runtime_manifest}
\alias{write_runtime_manifest}
\title{Write a runtime-manifest.json for the auto-download strategy}
\usage{
write_runtime_manifest(
  app_dir,
  app_type,
  platform,
  arch,
  config,
  verbose = TRUE
)
}
\description{
Write a runtime-manifest.json for the auto-download strategy
}
\keyword{internal}
