% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-templates.R
\name{bake_dockerfile_dependencies}
\alias{bake_dockerfile_dependencies}
\title{Append app-specific package installs to the Dockerfile}
\usage{
bake_dockerfile_dependencies(output_dir, dockerfile_dest)
}
\description{
Bakes the dependencies into the image at build time so container
launch doesn't have to compile/install packages on the user's machine.
}
\keyword{internal}
