% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-steps.R
\name{build_for_platforms}
\alias{build_for_platforms}
\title{Build for target platforms}
\usage{
build_for_platforms(output_dir, platform, arch, sign = FALSE, verbose = TRUE)
}
\arguments{
\item{output_dir}{Character Electron project directory}

\item{platform}{Character vector of target platforms}

\item{arch}{Character vector of target architectures}

\item{sign}{Logical whether to code-sign the build}

\item{verbose}{Logical whether to show progress}
}
\description{
Build for target platforms
}
\keyword{internal}
