% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-multi.R
\name{build_multi_app}
\alias{build_multi_app}
\title{Build multi-app Electron application}
\usage{
build_multi_app(
  apps_dir,
  output_dir,
  app_name,
  apps_manifest,
  default_type,
  runtime_strategy,
  sign,
  platform,
  arch,
  icon,
  config,
  overwrite,
  verbose,
  r_packages = NULL,
  r_repos = NULL,
  py_packages = NULL,
  py_index_urls = NULL
)
}
\description{
Build multi-app Electron application
}
\keyword{internal}
