% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{check_auto_update_status}
\alias{check_auto_update_status}
\title{Check Auto-Update Status}
\usage{
check_auto_update_status(appdir)
}
\arguments{
\item{appdir}{Character path to app directory}
}
\value{
Invisibly returns a list with update configuration.
}
\description{
Reports the current auto-update configuration status.
}
\examples{
\dontrun{
check_auto_update_status("path/to/app")
}

}
