% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-templates.R
\name{copy_and_bake_dockerfiles}
\alias{copy_and_bake_dockerfiles}
\title{Copy the Dockerfile for the container strategy and bake in app dependencies}
\usage{
copy_and_bake_dockerfiles(output_dir, app_type, verbose = TRUE)
}
\description{
Copy the Dockerfile for the container strategy and bake in app dependencies
}
\keyword{internal}
