% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_app_files}
\alias{copy_app_files}
\title{Copy application files to Electron project}
\usage{
copy_app_files(app_dir, output_dir, app_type, verbose = TRUE)
}
\arguments{
\item{app_dir}{Character source app directory}

\item{output_dir}{Character destination directory}

\item{app_type}{Character application type}

\item{verbose}{Logical whether to show progress}
}
\description{
Copy application files to Electron project
}
\keyword{internal}
