% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-templates.R
\name{copy_brand_assets}
\alias{copy_brand_assets}
\title{Copy branding assets (icon, splash image, tray icon) into the build}
\usage{
copy_brand_assets(output_dir, icon, config)
}
\description{
Copy branding assets (icon, splash image, tray icon) into the build
}
\keyword{internal}
