% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{disable_auto_updates}
\alias{disable_auto_updates}
\title{Disable Auto-Updates}
\usage{
disable_auto_updates(appdir, verbose = TRUE)
}
\arguments{
\item{appdir}{Character path to app directory}

\item{verbose}{Logical whether to show progress messages. Default \code{TRUE}.}
}
\value{
Invisibly returns the path to the updated config file.
}
\description{
Disables automatic update checking in the configuration file.
}
\examples{
\dontrun{
disable_auto_updates("path/to/app")
}

}
