% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_app}
\alias{example_app}
\title{Get Path to an Example Application}
\usage{
example_app(name)
}
\arguments{
\item{name}{Character string. Name of the example (see \code{\link[=available_examples]{available_examples()}}).}
}
\value{
Character string. Path to the example app directory.
}
\description{
Returns the path to a bundled example application directory. Use this
path as the \code{appdir} argument to \code{\link[=export]{export()}}.
}
\examples{
\dontrun{
# Get the R demo and export with any strategy
path <- example_app("r")
export(path, "output", app_type = "r-shiny", runtime_strategy = "system")
export(path, "output", app_type = "r-shinylive")

# Get the Python demo
path <- example_app("python")
export(path, "output", app_type = "py-shiny", runtime_strategy = "system")
}

}
