% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-multi.R
\name{export_multi_app}
\alias{export_multi_app}
\title{Export multi-app Shiny suite as Electron application}
\usage{
export_multi_app(
  appdir,
  destdir,
  config,
  runtime_strategy = NULL,
  sign = FALSE,
  platform = NULL,
  arch = NULL,
  icon = NULL,
  overwrite = FALSE,
  build = TRUE,
  run_after = FALSE,
  open_after = FALSE,
  verbose = TRUE
)
}
\description{
Export multi-app Shiny suite as Electron application
}
\keyword{internal}
