% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_npm_dependencies}
\alias{install_npm_dependencies}
\title{Install npm dependencies}
\usage{
install_npm_dependencies(output_dir, verbose = TRUE)
}
\arguments{
\item{output_dir}{Character Electron project directory}

\item{verbose}{Logical whether to show progress}
}
\description{
Install npm dependencies
}
\keyword{internal}
