% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_templates}
\alias{process_templates}
\title{Process and copy Electron templates}
\usage{
process_templates(output_dir, app_name, app_type, icon = NULL, verbose = TRUE)
}
\arguments{
\item{output_dir}{Character destination directory}

\item{app_name}{Character application name}

\item{app_type}{Character application type}

\item{icon}{Character path to icon file or NULL}

\item{verbose}{Logical whether to show progress}
}
\description{
Process and copy Electron templates
}
\keyword{internal}
