% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_app_name}
\alias{validate_app_name}
\title{Validate application name}
\usage{
validate_app_name(app_name)
}
\arguments{
\item{app_name}{Character application name}
}
\description{
Validate application name
}
\keyword{internal}
