% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_config}
\alias{validate_config}
\title{Validate configuration values}
\usage{
validate_config(config)
}
\arguments{
\item{config}{List of configuration values}
}
\value{
List of validated configuration
}
\description{
Checks configuration values and warns about invalid entries.
}
\keyword{internal}
