% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_multi_app_config}
\alias{validate_multi_app_config}
\title{Validate multi-app configuration}
\usage{
validate_multi_app_config(config, basedir)
}
\arguments{
\item{config}{List. Configuration with apps array.}

\item{basedir}{Character. Base directory for resolving relative paths.}
}
\description{
Validate multi-app configuration
}
\keyword{internal}
