% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-r.R
\name{validate_r_available}
\alias{validate_r_available}
\title{Validate R is available on the system}
\usage{
validate_r_available()
}
\value{
Invisible character string with the path to Rscript.
}
\description{
Checks that Rscript can be found and executed. Used by the "system"
runtime strategy where the end user must have R installed.
}
\keyword{internal}
