% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_shinylive_output}
\alias{validate_shinylive_output}
\title{Validate shinylive output}
\usage{
validate_shinylive_output(output_dir)
}
\arguments{
\item{output_dir}{Character path to shinylive output}
}
\description{
Validate shinylive output
}
\keyword{internal}
